Added files via upload. First BETA release.
liat  is  an  ncurses application that reads lines from FILE or stdin and prints the lines to the screen. The FILE is monitored and as the FILE grows the appended data is written to the screen. The top line of the screen is updated and the rest of the screen is scrolled down. The bottom line is lost each time the screen is full and the top line is updated.